Fixed breaking images. Refactored CSS in common sections into shared folder
